packages/db_tests/pyproject.toml
[tool.poetry]
name = "db_tests"
version = "0.1.0"
description = ""
authors = ["Piotr Styczyński <piotr@styczynski.in>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
"testing.postgresql" = "^1.3.0"
SQLAlchemy = "^1.4.37"
psycopg2 = "^2.9.3"
pgsanity = "^0.2.9"
rich = "^12.4.4"
pydantic = "^1.9.1"
tqdm = "^4.64.0"
pandas = "^1.4.3"
parametrized = "^0.1"
pytest-diff = "^0.1.14"
pytest-rich = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
